refactor: 修改slotComponent为ChunkComponent, 去除subComponent返回类型中的alldatas和events字段
