+++ src/api/ChunkComponent/ChunkComputed/ChunkComputedConstraint.ts
