+++ src/api/ChunkComponent/ChunkData/chunkDataConstraint.ts
