+++ src/api/SubComponent/SubReturnType/test/normal.test.ts
@@ -37,9 +37,9 @@ type SubDocExpect = {
     captrueComposed: number | CaptureComposed;
     bubblesCaptrueComposed: number | BubblesCaptureComposed;
   };
-  allDatas: {
-    aaa_str: "string";
-  };
+  // allDatas: {
+  //   aaa_str: "string";
+  // };
 };
 
 Checking<typeof SubDoc, SubDocExpect, Test.Pass>();
@@ -59,18 +59,18 @@ const removeCatchedEvents = SubComponent<{}, CompDoc>()({
 });
 void removeCatchedEvents;
 // 返回穿透的自定义事件,并去除了前缀
-type RemoveSubDoc = {
-  allDatas: {
-    aaa_str: "string";
-  };
-  events: {
-    aaa_bubblesCaptrueComposed_catch(): void;
-    aaa_bubblesComposed_catch(): void;
-    aaa_captrueComposed_catch(): void;
-  };
-};
+// type RemoveSubDoc = {
+//   allDatas: {
+//     aaa_str: "string";
+//   };
+//   events: {
+//     aaa_bubblesCaptrueComposed_catch(): void;
+//     aaa_bubblesComposed_catch(): void;
+//     aaa_captrueComposed_catch(): void;
+//   };
+// };
 
-Checking<typeof removeCatchedEvents, RemoveSubDoc, Test.Pass>();
+Checking<typeof removeCatchedEvents, never, Test.Pass>();
 // 去除有后缀时,已经声明的穿透事件
 const removeCatchedEventsOfSuffix = SubComponent<{}, CompDoc, "aa">()({
   data: {
@@ -84,15 +84,15 @@ const removeCatchedEventsOfSuffix = SubComponent<{}, CompDoc, "aa">()({
 });
 void removeCatchedEventsOfSuffix;
 // 返回穿透的自定义事件,并去除了前缀
-type RemoveCatchedEventsOfSuffix = {
-  allDatas: {
-    aaaAa_str: "string";
-  };
-  events: {
-    aaaAa_bubblesCaptrueComposed_catch(): void;
-    aaaAa_bubblesComposed_catch(): void;
-    aaaAa_captrueComposed_catch(): void;
-  };
-};
+// type RemoveCatchedEventsOfSuffix = {
+//   allDatas: {
+//     aaaAa_str: "string";
+//   };
+//   events: {
+//     aaaAa_bubblesCaptrueComposed_catch(): void;
+//     aaaAa_bubblesComposed_catch(): void;
+//     aaaAa_captrueComposed_catch(): void;
+//   };
+// };
 
-Checking<typeof removeCatchedEventsOfSuffix, RemoveCatchedEventsOfSuffix, Test.Pass>();
+Checking<typeof removeCatchedEventsOfSuffix, never, Test.Pass>();
