+++ src/index.ts
@@ -1,7 +1,7 @@
+import { ChunkComponent } from "./api/ChunkComponent";
 import { DefineComponent } from "./api/DefineComponent";
 import { type IInjectInfo, instanceConfig } from "./api/InstanceInject/instanceConfig";
 import { RootComponent } from "./api/RootComponent";
-import { ChunkComponent } from "./api/SlotComponent";
 import { SubComponent } from "./api/SubComponent";
 import { navigateBack, navigateTo, redirectTo } from "./api/wxSugar";
 import type { CreateComponentType } from "./types/CreateComponentType";
